+++ src/main.rs
@@ -1,7 +1,7 @@
 use std::env;
 
 fn main() {
-    match dinero::run_app(env::args().collect()) {
+    match dinero::run_application(env::args().collect()) {
         Ok(_) => {}
         Err(_) => std::process::exit(1),
     }