+++ src/parser.rs
@@ -16,9 +16,9 @@ use pest::Parser;
 
 mod include;
 // pub mod tokenizers;
+pub mod expressions;
 pub mod tokenizers;
 pub(crate) mod utils;
-pub mod value_expr;
 
 use tokenizers::transaction;