+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 Changelog file for dinero-rs project, a command line application for managing finances.
 ## [0.26.0] - xxx
+### Added
+- Support for custom format strings
+### Changed
+- Now the reports are parsed using a format string, which makes them more flexible
 ## [0.25.0] - 2021-03-31
 ### Added
 - nicer error reporting